requirements.txt
flask
flask_sqlalchemy
opencv-python
numpy
scikit-image
pillow
reportlab